refresh 6502 core

+++ cpu.h
@@ -22,7 +22,7 @@ class CPU {
 public:
 	virtual void run(unsigned instructions) =0;
 	virtual void reset() =0;
-	virtual void raise(int level) =0;
+	virtual void raise(uint8_t level) =0;
 	virtual char *status(char *buf, size_t n, bool hdr) =0;
 
 	typedef CPU *(*buildfn) (Memory &);
